🚧 Layouts types extractions
